+++ lib/utils.js
@@ -7,6 +7,7 @@ const xml2js = require('xml2js-es6-promise');
 const uuidv1 = require('uuid/v1');
 
 const attachment = require('./attachments');
+const { checkRateLimitParams, rateLimit } = require('./helpers');
 
 const HTTP_ERROR_CODE_REBOUND = new Set([408, 423, 429, 500, 502, 503, 504]);
 
@@ -84,6 +85,8 @@ module.exports.processMethod = async function (msg, cfg) {
       `Method "${method}" isn't one of the: ${Object.keys(methodsMap)}.`,
     );
   }
+
+  checkRateLimitParams(cfg, emitter.logger);
 
   /*
    if cfg.followRedirect has value doNotFollowRedirects
@@ -181,7 +184,7 @@ module.exports.processMethod = async function (msg, cfg) {
         output.attachments = msg.attachments;
         await emitter.emit('data', output);
       }
-      await rateLimit(cfg);
+      await rateLimit(cfg, emitter.logger);
       await emitter.emit('end');
     })
     .catch(await buildErrorStructure);
@@ -371,12 +374,12 @@ module.exports.processMethod = async function (msg, cfg) {
       };
       emitter.logger.debug('Component output: %o', output);
       await emitter.emit('data', messages.newMessageWithBody(output));
-      await rateLimit(cfg);
+      await rateLimit(cfg, emitter.logger);
       await emitter.emit('end');
     } else {
       emitter.logger.error('Component error: %o', e);
       await emitter.emit('error', e);
-      await rateLimit(cfg);
+      await rateLimit(cfg, emitter.logger);
       await emitter.emit('end');
     }
   }
@@ -541,20 +544,20 @@ module.exports.processMethod = async function (msg, cfg) {
     }
   }
 
-  function sleep(delay) {
-    return new Promise((resolve) => {
-      setTimeout(resolve, delay * 1000);
-    });
-  }
+  // function sleep(delay) {
+  //   return new Promise((resolve) => {
+  //     setTimeout(resolve, delay * 1000);
+  //   });
+  // }
 
-  async function rateLimit(cfg) {
-    const { delayBetweenCalls } = cfg;
-    if (delayBetweenCalls) {
-      const delay = parseInt(delayBetweenCalls, 10);
-      emitter.logger.info('Delay Between Calls is set to:', delay);
-      emitter.logger.debug('Delay is start', new Date());
-      await sleep(delay);
-      emitter.logger.debug('Delay is done', new Date());
-    }
-  }
+  // async function rateLimit(cfg) {
+  //   const { delayBetweenCalls } = cfg;
+  //   if (delayBetweenCalls) {
+  //     const delay = parseInt(delayBetweenCalls, 10);
+  //     emitter.logger.info('Delay Between Calls is set to:', delay);
+  //     emitter.logger.debug('Delay is start', new Date());
+  //     await sleep(delay);
+  //     emitter.logger.debug('Delay is done', new Date());
+  //   }
+  // }
 };
